1、公告管理相关接口 2、公告管理表定义 3、修正删除操作记录更新人、更新日期

+++ back-end/color-system/src/main/java/com/colorone/system/domain/entity/BaseMenu.java
@@ -26,7 +26,7 @@ public class BaseMenu extends BaseEntity {
      * comment '菜单id'   bigint(10)   not null autoIncrement
      */
     @TableId(value = "menu_id", type = IdType.AUTO)
-    Integer menuId;
+    Long menuId;
 
     /**
      * comment '菜单接口名称' varchar(50) not null
